requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiohttp==3.9.5
    # via discord-py
aiosignal==1.3.1
    # via aiohttp
attrs==23.2.0
    # via aiohttp
discord-py==2.3.2
    # via -r requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
idna==3.7
    # via yarl
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
python-dateutil==2.9.0.post0
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
yarl==1.9.4
    # via aiohttp
